requirements.txt
numpy
matplotlib
pandas
scipy
librosa
soundfile
Pillow
